[IMPR] Create utililty
