+++ app.py
@@ -11,10 +11,25 @@
 
 from flask import Flask
 from flask import render_template
+from flask import request
+import utils
 
 app = Flask(__name__)
 
 
 @app.route("/")
+@app.route("/", methods = ["GET", "POST"])
 def accueil():
-    return render_template('index.html')
+
+	# If form was submitted 
+	if request.methods == "POST":
+		limit = 10
+		if request.form['limit'] != "":
+    		limit = request.form['limit']
+		
+		articles = utils.getArticles(limit)
+		return render_template('index.html', articles=articles)
+    
+    # If request is GET
+    else:
+    	return render_template('index.html')
