+++ templates/login.html
@@ -40,24 +40,6 @@
     <button class="btn btn-lg btn-primary btn-block" type="submit">Afficher</button>
   </form>
 </div>
-<div>
-  <table class="table">
-  <thead class="thead-dark">
-    <tr>
-      <th scope="col">#</th>
-      <th scope="col">Title</th>
-      <th scope="col">Templates</th>
-    </tr>
-  </thead>
-  <tbody>
-    <tr>
-      <th scope="row">1</th>
-      <td>Mark</td>
-      <td>Otto</td>
-    </tr>
-  </tbody>
-</table>
-</div>
 </main>
 <br>
 <br>
