+++ update.py
@@ -8,6 +8,7 @@ import json
 import os
 import requests
 from tinydb import TinyDB
+import utils
 
 
 def app():
@@ -59,7 +60,7 @@ def app():
 						"templates" : page_templates
 					}
 
-					createArticle(article)
+					utils.createArticle(article)
 					print(page_title + " was saved in the DB.")
 					
 					articles_count += 1
@@ -74,32 +75,6 @@ def runMediaWikiRequest(url):
 	resultatJson = requests.get(url).content
 	# return the json text
 	return resultatJson
-
-def getDb():
-	""" Return a Database (TinyDB) object """
-	dbFolder = "database"
-	if not os.path.exists(dbFolder):
-    	os.makedirs(dbFolder)
-
-	db = TinyDB(dbFolder + '/db.json')
-	return db
-
-
-def getArticles(limit=10):
-	""" Return all articles in the database """ 
-	db = getDb()
-	results = db.all()
-
-	if len(results) < limit:
-		return results[0:limit-1]
-	
-	return results
-
-
-def createArticle(article):
-	""" Insert new article database """ 
-	db = getDb()
-	return db.insert(article)
 
 
 def printNotice(notice):
